apps/crawler/pyproject.toml
[project]
name = "jobseek-crawler"
version = "0.1.0"
description = "Job listing monitor and extraction crawler"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30",
    "click>=8.1",
    "httpx>=0.28",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0",
    "structlog>=24.4",
    "python-dotenv>=1.0",
    "polars>=1.38.1",
]

[project.scripts]
scheduler = "src.scheduler:main"
ws = "src.workspace.cli:main"

[dependency-groups]
dev = [
    "playwright>=1.58.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.15.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
